added modal to create new projects
