+++ src/index.js
@@ -1,8 +1,10 @@
 import {Task} from './task';
+import {cont, addTask} from './addtask';
 
 export let task = [];
- let cont =document.getElementById('conten');
+export let project = [];
  const createButton = document.getElementById('create');
+ const pcreateButton = document.getElementById('pcreate');
  
  createButton.onclick = function create () {
   const title = document.getElementById('orangeForm-title').value;
@@ -15,36 +17,15 @@ export let task = [];
   task.push(newTask);
   addTask(title, description, dueDate, priority);
  }
-
-function addTask(title, description, dueDate, priority) {
-  const box = document.createElement('div');
-  box.className = 'col-3 task';
-  cont.appendChild(box);
-  const card = document.createElement('div');
-  card.className = 'card';
-  box.appendChild(card);
-  const del = document.createElement('a');
-  del.className = 'btn bbd';
-  del.appendChild(document.createTextNode('X'));
-  card.appendChild(del);
-  const cardBody = document.createElement('div');
-  cardBody.className = 'card-body';
-  card.appendChild(cardBody);
-  const cardTitle = document.createElement('h3');
-  cardTitle.className = 'card-title';
-  cardTitle.appendChild(document.createTextNode(title));
-  cardBody.appendChild(cardTitle);
-  const bookDescription = document.createElement('p');
-  bookDescription.className = 'card-text';
-  bookDescription.appendChild(document.createTextNode(`DUE DATE: ${dueDate}`));
-  cardBody.appendChild(bookDescription);
-  const numPage = document.createElement('h4');
-  numPage.className = 'card-title';
-  numPage.appendChild(document.createTextNode(` PRIORITY: ${priority}`));
-  cardBody.appendChild(numPage);
-}
+ 
+  pcreateButton.onclick = function Pcreate () {
+  const name = document.getElementById('orangeForm-title').value;
+  console.log("project");
+ 
+ }
 
 const rendertask =() => {
+    cont.innerHTML='';
   for (let i = 0; i <task.length; i++) {
   const { title } = task[i];
   const { description } = task[i];
@@ -54,8 +35,9 @@ const rendertask =() => {
 }
 }
 
- rendertask();
-
+if (task.length!=0){
+    rendertask();
+}
 // render(title, description, dueDate, priority);
 // render(title, description, dueDate, priority);
 // render(title, description, dueDate, priority);
